api/requirements.txt
flask
flask_cors
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.1
prometheus-flask-exporter==0.23.2
pika==1.3.2
